Move libraryUtils back to utils/common
