+++ utils/common/libraryUtils.ts
