+++ utils/test/index.ts
@@ -32,6 +32,3 @@ export {
   mineBlockAsync,
   cacheBeforeEach
 } from "./testingUtils";
-export {
-  convertLibraryNameToLinkId
-} from "./libraryUtils";
